Extracted Menu Options to option package
