+++ app/src/main/java/com/example/deliveryapp/ui/components/menu/option/OptionSection.kt
@@ -1,0 +1,34 @@
+package com.example.deliveryapp.ui.components.menu.option
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun OptionSection() {
+    Column(
+        Modifier.fillMaxWidth()
+    ) {
+        Column(
+            verticalArrangement = Arrangement.spacedBy(16.dp),
+            modifier = Modifier
+                .background(Color.White)
+                .fillMaxWidth()
+                .padding(16.dp)
+        ) {
+            Text("Upgrade to Combo", fontWeight = FontWeight.Bold, fontSize = 18.sp)
+            Option("Fries & Pepsi")
+            Option("Fries & Diet Pepsi")
+            Option("Fries & 7 Up")
+        }
+    }
+}
